move relavant components into dashboard dir
