+++ packages/dashboard/src/components/dashboard/commands-panel.tsx
@@ -17,7 +17,7 @@ import {
   LoopRequestForm,
 } from 'react-components';
 import { v4 as uuidv4 } from 'uuid';
-import { ResourcesContext } from './app-contexts';
+import { ResourcesContext } from '../app-contexts';
 
 const debug = Debug('OmniPanel:CommandsPanel');
