+++ packages/dashboard/src/components/dashboard/tests/commands-panel.test.tsx
@@ -2,10 +2,10 @@ import { createMount } from '@material-ui/core/test-utils';
 import * as RomiCore from '@osrf/romi-js-core-interfaces';
 import React from 'react';
 import { DeliveryRequestForm, LoopRequestForm } from 'react-components';
-import ResourceManager from '../../managers/resource-manager';
-import fakeFleets from '../../mock/data/fleets';
-import fakeResources from '../../mock/data/resources';
-import { ResourcesContext } from '../app-contexts';
+import ResourceManager from '../../../managers/resource-manager';
+import fakeFleets from '../../../mock/data/fleets';
+import fakeResources from '../../../mock/data/resources';
+import { ResourcesContext } from '../../app-contexts';
 import CommandsPanel from '../commands-panel';
 
 const mount = createMount();
