+++ packages/dashboard/src/components/dashboard/tests/dashboard.test.tsx
@@ -2,12 +2,12 @@ import { createMount } from '@material-ui/core/test-utils';
 import { ReactWrapper } from 'enzyme';
 import React from 'react';
 import ReactTestUtils from 'react-dom/test-utils';
-import Dashboard from '../dashboard/dashboard';
+import Dashboard from '../dashboard';
 
 const mount = createMount();
 
 // react-leaflet doesn't work well in jsdom.
-jest.mock('./../schedule-visualizer', () => () => null);
+jest.mock('./../../schedule-visualizer', () => () => null);
 
 it('renders without crashing', async () => {
   URL.createObjectURL = jest.fn();
