+++ packages/dashboard/src/components/dashboard/tests/task-summary-panel.test.tsx
@@ -1,8 +1,8 @@
-import React from 'react';
 import { render } from '@testing-library/react';
 import userEvent from '@testing-library/user-event';
+import React from 'react';
+import fakeTaskSummary from '../../../mock/data/task-summary';
 import TaskSummaryPanel from '../task-summary-panel';
-import fakeTaskSummary from '../../mock/data/task-summary';
 
 describe('It renders correctly', () => {
   test('Renders tree items', () => {
