+++ packages/dashboard/src/components/dashboard/tour/tour-navigation-control.tsx
