--- packages/dashboard/src/components/tests/main-menu.test.tsx
@@ -1,9 +1,0 @@
-import React from 'react';
-import ReactDOM from 'react-dom';
-import MainMenu from '../main-menu';
-
-it('renders without crashing', () => {
-  const div = document.createElement('div');
-  ReactDOM.render(<MainMenu />, div);
-  ReactDOM.unmountComponentAtNode(div);
-});
